FW-004 added timer animation [SL,LH]
