+++ app/src/main/res/drawable/circle_progress.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+<item>
+    <shape
+        android:shape="ring"
+        android:thicknessRatio="20"
+        android:useLevel="false">
+        <solid android:color="#DDD" />
+    </shape>
+</item>
+<item>
+    <rotate
+        android:fromDegrees="270"
+        android:toDegrees="270">
+        <shape
+            android:shape="ring"
+            android:thicknessRatio="20"
+            android:useLevel="true">
+            <solid android:color="#080" />
+        </shape>
+    </rotate>
+</item>
+</layer-list>
